Refactoring the components parsing functions to extract them to reuse them.
